+++ src/library/zoid/modal/component.js
@@ -136,11 +136,11 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                 required: false,
                 value: validate.ecToken
             },
-            contextualComponent: {
+            contextualComponents: {
                 type: 'string',
-                queryParam: 'contextual_component',
+                queryParam: 'contextual_components',
                 required: false,
-                value: validate.contextualComponent
+                value: validate.contextualComponents
             },
 
             // Callbacks
@@ -442,7 +442,7 @@ export default createGlobalVariableGetter('__paypal_credit_modal__', () =>
                 type: 'string',
                 queryParam: 'features',
                 required: false,
-                value: getFeatures
+                value: validate.features ?? getFeatures
             }
         }
     })